Whoo, lots of progress today. Still working on implementing the control message stuff at the receiver.  Next focus: hook the connection-related messages into the CMMSocketImpl data structures.
git-svn-id: file:///home/svnshare/juggler_2.6.22.14/libcmm/trunk@370 1971e5c1-a347-0410-b49b-82492f0cf60c

+++ irob.h
@@ -27,10 +27,12 @@
  * longer needed as separate functions.
  */
 
+#include <sys/types.h>
+
 /* Identifer for an IROB, assigned by begin_irob[_label] 
  * and released by end_irob 
  */
-typedef int irob_id_t;
+typedef u_long irob_id_t;
 
 /* Begin an IROB and return the new IROB id.
  * If deps is non-NULL, it should point to a depnum-sized array of
